Add files via upload
The i1d3util tool can read the data out of the iD3 and write it to disk, it can also take data on disk and write it back into the i1D3.  
By doing this, you can backup and restore your probe.
For example, if you have a retail probe and the signature data from an oem probe, you can load the oem signature into the retail probe.  
The probe will now operate as if it were a factory oem probe.